buses-backend/requirements.txt
Django==3.1.2
djangorestframework==3.11.1
django-cors-headers==3.6.0
drf-nested-routers==0.92.1
gunicorn==20.0.4
uvicorn[standard]==0.12.2
psycopg2==2.8.6